+++ packages/yume-engine/src/index.ts
@@ -63,6 +63,7 @@ export * from './save-codecs/index.js';
 export * from './fs/index.js';
 export * from './process/index.js';
 export * from './bundle/index.js';
+export * from './icon/index.js';
 export * from './utils/index.js';
 
 const MACHO_MAGICS = new Set<number>([